Samples/python/requirements.txt
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.25.3, < 2.1.0
pydantic >= 2
typing-extensions >= 4.7.1
smabbler-api-client >= 2.1.2
pandas >= 2.1.4
scikit-learn >= 1.3.2
